requirements.txt
pillow
djangorestframework
collections
django-cors-headers